third

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
 
-    public float calculate(int kilo, float metres) {
+    public int calculate(int kilo, float metres) {
 
         float index;
